+++ tests/HttpRequestCurlTest.php
@@ -4,8 +4,8 @@ declare(strict_types=1);
 
 namespace BEAR\Resource;
 
-use BEAR\Dev\Http\BuiltinServer;
 use BEAR\Resource\Exception\HttpRequestException;
+use Koriym\PhpServer\PhpServer;
 use PHPUnit\Framework\TestCase;
 use RuntimeException;
 
@@ -16,11 +16,11 @@ final class HttpRequestCurlTest extends TestCase
     private const HOST = '127.0.0.1:8101';
     private const URL = 'http://127.0.0.1:8101/';
 
-    private static BuiltinServer $server;
+    private static PhpServer $server;
 
     public static function setUpBeforeClass(): void
     {
-        self::$server = new BuiltinServer(self::HOST, __DIR__ . '/Server/http_request_curl.php');
+        self::$server = new PhpServer(self::HOST, __DIR__ . '/Server/http_request_curl.php');
         self::$server->start();
     }